mysite/requirements.txt
astroid==2.0.4
Django==2.1.3
django-multiselectfield==0.1.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.1.1
pytz==2018.7
six==1.11.0
wrapt==1.10.11
